[Bronze III] Title: 삼각 김밥, Time: 124 ms, Memory: 14232 KB -BaekjoonHub

+++ 백준/Bronze/2783. 삼각 김밥/삼각 김밥.java
@@ -18,7 +18,8 @@ public class Main {
             y = Integer.parseInt(st.nextToken());
             min = Math.min((1000.0 / y) * x, min);
         }
-        System.out.println((double) Math.round(min * 100) / 100.0);
+//        System.out.println((double) Math.round(min * 100) / 100.0);
+        System.out.println(Math.round(min * 100) / 100.0);
 //        System.out.printf("%.2f", min);
     }
 }
